Cambiamos ubicación de los DTOs y re-configuramos el FluentValidation
