--- Model/Entities/DTOs/StadiumDto.cs
@@ -1,15 +1,0 @@
-﻿using Model.Entities;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Model.Entities.DTOs
-{
-    public class StadiumDto
-    {
-        // Nombre del Estadio
-        public string Name { get; set; }
-    }
-}
