+++ NetWebApi/DTOs/StandingDto.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Model.Entities.DTOs
+namespace NetWebApi.DTOs
 {
     public class StandingDto
     {
