--- Repository/Configuration/AutoMapperConfig.cs
@@ -1,26 +1,0 @@
-﻿using AutoMapper;
-using Model.Entities;
-using Model.Entities.DTOs;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Repository.Configuration
-{
-    public class AutoMapperConfig : Profile
-    {
-        public AutoMapperConfig()
-        {
-            // AutoMapper Club
-            CreateMap<Club, ClubDto>()
-                .ForMember(dest => dest.Name, org => org.MapFrom(src => src.Name));
-            CreateMap<ClubDto, Club>();
-            CreateMap<ClubPostDto, Club>();
-
-            // AutoMaper Player
-            CreateMap<PlayerPostDto, Player>();
-        }
-    }
-}
